+++ database/migrations/gateway/2026_01_05_031642_create_callback_logs_table.php
@@ -4,21 +4,22 @@ use Illuminate\Database\Migrations\Migration;
 use Illuminate\Database\Schema\Blueprint;
 use Illuminate\Support\Facades\Schema;
 
-return new class extends Migration
-{
+return new class extends Migration {
     /**
      * Run the migrations.
      */
     public function up(): void
     {
-        Schema::connection('gateway')->create('callback_attempts', function (Blueprint $table) {
-            $table->id();
-            $table->foreignId('transaction_id')->constrained('transactions')->onDelete('cascade');
-            $table->string('callback_url');
-            $table->integer('response_code')->nullable();
-            $table->string('status'); // SUCCESS / FAILED
-            $table->timestamp('created_at')->useCurrent();
-        });
+        if (!Schema::connection('gateway')->hasTable('callback_attempts')) {
+            Schema::connection('gateway')->create('callback_attempts', function (Blueprint $table) {
+                $table->id();
+                $table->foreignId('transaction_id')->constrained('transactions')->onDelete('cascade');
+                $table->string('callback_url');
+                $table->integer('response_code')->nullable();
+                $table->string('status'); // SUCCESS / FAILED
+                $table->timestamp('created_at')->useCurrent();
+            });
+        }
     }
 
     /**
